Remove wording about writing large files to butler
I don't think it's necessary to tell people to use the butler to put
data.  If they are using the butler, they will naturally put things
back in the butler.  If they aren't, we don't want to pay the support
overhead to teach them to put things in the butler.